Level "Welcome to Antarctica": Update and extend billboards.
The billboard and infoblock text for the "reset point" explanation has been
corrected. This resolves bug 591.

A billboard explaining the butt-jump and that wooden boxes can be destroyed
this way has been added. A couple of coins have been added to the level that
can only be reached using this technique.

SVN-Revision: 6669